modified a lot
