+++ WindAi/deep_learning/optuna_trials_results.csv
@@ -1,0 +1,31 @@
+number,value,datetime_start,datetime_complete,duration,params_dense_units,params_gru_units,params_l2_strength,params_learning_rate,params_loss,params_n_layers,state
+0,1.784064769744873,2025-07-29 13:45:05.944581,2025-07-29 14:03:50.567084,0 days 00:18:44.622503,32,64,1.6942442111777894e-05,0.0008591929770764513,mse,4,COMPLETE
+1,1.009299874305725,2025-07-29 14:03:50.567340,2025-07-29 15:10:59.410339,0 days 01:07:08.842999,8,512,4.354069879152364e-05,0.0002755254820239147,mae,7,COMPLETE
+2,2.019468307495117,2025-07-29 15:10:59.410659,2025-07-29 15:25:09.144471,0 days 00:14:09.733812,16,64,0.00022066138380395398,0.00032250467820366536,mse,3,COMPLETE
+3,2.030259609222412,2025-07-29 15:25:09.144798,2025-07-29 15:52:20.965214,0 days 00:27:11.820416,32,64,0.000289211330740887,0.00020927144617204183,mse,6,COMPLETE
+4,1.6245245933532715,2025-07-29 15:52:20.965536,2025-07-29 16:17:53.562095,0 days 00:25:32.596559,32,256,1.905523039881056e-05,0.00018083309770826803,mse,4,COMPLETE
+5,0.9596647024154663,2025-07-29 16:17:53.562349,2025-07-29 16:30:54.287857,0 days 00:13:00.725508,16,256,5.985387351673873e-05,0.0006742270378388501,mae,2,COMPLETE
+6,2.0262410640716553,2025-07-29 16:30:54.288115,2025-07-29 16:40:53.813920,0 days 00:09:59.525805,32,64,1.1330414839452644e-05,0.00012166333167779617,mse,2,COMPLETE
+7,1.586401343345642,2025-07-29 16:40:53.814256,2025-07-29 17:06:28.315969,0 days 00:25:34.501713,32,256,9.125829258224448e-05,0.0002765539388883748,mse,4,COMPLETE
+8,1.060531497001648,2025-07-29 17:06:28.316349,2025-07-29 17:22:04.856731,0 days 00:15:36.540382,32,128,5.098536939592102e-05,0.00021545585835872614,mae,3,COMPLETE
+9,1.577290654182434,2025-07-29 17:22:04.857065,2025-07-29 18:06:29.508794,0 days 00:44:24.651729,16,256,0.0003469048540456968,0.0007348041026141512,mse,7,COMPLETE
+10,1.1352717876434326,2025-07-29 18:06:29.509122,2025-07-29 18:14:48.937603,0 days 00:08:19.428481,16,32,0.0009442946377101649,0.000528326918602806,mae,2,COMPLETE
+11,0.9620542526245117,2025-07-29 18:14:48.937994,2025-07-29 19:12:03.088139,0 days 00:57:14.150145,8,512,4.3606327501832405e-05,0.0004886482312469708,mae,6,COMPLETE
+12,1.0079208612442017,2025-07-29 19:12:03.088466,2025-07-29 20:09:17.293029,0 days 00:57:14.204563,8,512,4.318285701293864e-05,0.00048757895041894094,mae,6,COMPLETE
+13,0.9757422804832458,2025-07-29 20:09:17.293358,2025-07-29 20:57:24.865734,0 days 00:48:07.572376,8,512,0.00010248463374638506,0.0005178642477196594,mae,5,COMPLETE
+14,1.0851950645446777,2025-07-29 20:57:24.865926,2025-07-29 21:18:32.791018,0 days 00:21:07.925092,8,32,9.5217781760029e-05,0.00042403926105410045,mae,6,COMPLETE
+15,0.9978026747703552,2025-07-29 21:18:32.791340,2025-07-29 21:43:28.656705,0 days 00:24:55.865365,16,128,2.678260251462557e-05,0.0006835409376821271,mae,5,COMPLETE
+16,0.9896655678749084,2025-07-29 21:43:28.657063,2025-07-29 22:12:03.654755,0 days 00:28:34.997692,16,512,0.00016835917689806142,0.000994694912325606,mae,3,COMPLETE
+17,1.0084302425384521,2025-07-29 22:12:03.655098,2025-07-29 22:43:53.147704,0 days 00:31:49.492606,8,256,6.030738084147544e-05,0.0006477273499912471,mae,5,COMPLETE
+18,0.9838552474975586,2025-07-29 22:43:53.148111,2025-07-29 23:41:41.206053,0 days 00:57:48.057942,16,512,2.6145499149216607e-05,0.00038073566815034754,mae,6,COMPLETE
+19,0.9645217061042786,2025-07-29 23:41:41.206376,2025-07-29 23:54:45.936499,0 days 00:13:04.730123,8,256,0.00014598002759636192,0.0006304037681749362,mae,2,COMPLETE
+20,1.2238473892211914,2025-07-29 23:54:45.936682,2025-07-30 00:19:07.889054,0 days 00:24:21.952372,16,32,0.0005578767193680818,0.0003861919330626342,mae,7,COMPLETE
+21,0.959435760974884,2025-07-30 00:19:07.889386,2025-07-30 00:32:13.707469,0 days 00:13:05.818083,8,256,0.00015048431166904874,0.0006211186994392116,mae,2,COMPLETE
+22,0.9612458944320679,2025-07-30 00:32:13.707802,2025-07-30 00:51:37.785153,0 days 00:19:24.077351,8,256,7.28945784424765e-05,0.000814195835551651,mae,3,COMPLETE
+23,0.9763323068618774,2025-07-30 00:51:37.785485,2025-07-30 01:04:46.350586,0 days 00:13:08.565101,8,256,6.811056298153404e-05,0.0009981435842035219,mae,2,COMPLETE
+24,1.0040913820266724,2025-07-30 01:04:46.350951,2025-07-30 01:24:11.552173,0 days 00:19:25.201222,8,256,0.00014362627986066735,0.0008438882547523694,mae,3,COMPLETE
+25,1.025107502937317,2025-07-30 01:24:11.552510,2025-07-30 01:37:23.912068,0 days 00:13:12.359558,8,256,7.646795718784439e-05,0.0006054369321424459,mae,2,COMPLETE
+26,0.990962564945221,2025-07-30 01:37:23.912410,2025-07-30 01:56:45.112330,0 days 00:19:21.199920,8,256,0.00012288830715446157,0.0007897442525902223,mae,3,COMPLETE
+27,0.9622129797935486,2025-07-30 01:56:45.112658,2025-07-30 02:09:50.403412,0 days 00:13:05.290754,16,256,0.00020439247529317113,0.000589022045567814,mae,2,COMPLETE
+28,0.975712776184082,2025-07-30 02:09:50.403753,2025-07-30 02:29:16.354773,0 days 00:19:25.951020,8,256,0.00039848056609507006,0.0007919411326293256,mae,3,COMPLETE
+29,0.9777278900146484,2025-07-30 02:29:16.354968,2025-07-30 02:49:42.582252,0 days 00:20:26.227284,16,128,3.0404570568311235e-05,0.0008699495058735019,mae,4,COMPLETE
