+++ project.sh
@@ -1,20 +1,30 @@
 #!/bin/bash
-#SBATCH --time=02:20:00         
+#SBATCH --time=16:00:00         
 #SBATCH --partition=gpu         
-#SBATCH --gres=gpu:1            
-#SBATCH --mem=32000              
+#SBATCH --gres=gpu:2            
+#SBATCH --mem=32000
 
-# Load required modules
+module spider CUDA
+module spider cuDNN
+module purge
 module load Python/3.10.4-GCCcore-11.3.0
-module load CUDA/11.7.0
-module load Boost/1.79.0-GCC-11.3.0
+module load CUDA/11.8.0
 
-# (Optional) Activate a virtual environment if needed
+export CUDA_HOME=$EBROOTCUDA
+export XLA_FLAGS=--xla_gpu_cuda_data_dir=$CUDA_HOME
+
+
+module load cuDNN/8.7.0.84-CUDA-11.8.0
+
+
+# Activate your virtual environment
 source /home2/s5549329/windAI_rug/venv/bin/activate
 
-python --version
+echo "Running on GPU node..."
+python -c "import tensorflow as tf; print('TF version:', tf.__version__); print('GPU:', tf.config.list_physical_devices('GPU')); print('CUDA version:', tf.sysconfig.get_build_info().get('cuda_version')); print('cuDNN version:', tf.sysconfig.get_build_info().get('cudnn_version'))"
 
-# Run your Python project
-python /home2/s5549329/windAI_rug/WindAi/deep_learning/models/RNN.py
+# Run your Python script
+python /home2/s5549329/windAI_rug/WindAi/deep_learning/models/GRU_main.py
 
+# Deactivate environment
 deactivate
